+++ lib/get_rx/src/rx_types/rx_core/rx_string.dart
@@ -21,23 +21,8 @@ class RxString extends Rx<String> implements Comparable<String>, Pattern {
 }
 
 /// Rx class for nullable `String` type.
-class RxnString extends Rx<String?> implements Comparable<String>, Pattern {
+class RxnString extends Rx<String?> {
   RxnString([super.initial]);
-
-  @override
-  Iterable<Match> allMatches(String string, [int start = 0]) {
-    return value!.allMatches(string, start);
-  }
-
-  @override
-  Match? matchAsPrefix(String string, [int start = 0]) {
-    return value!.matchAsPrefix(string, start);
-  }
-
-  @override
-  int compareTo(String other) {
-    return value!.compareTo(other);
-  }
 }
 
 /// Extension on [Rx<String>] providing standard operators.