+++ lib/get_navigation/src/routes/core/get_navigation_interface.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/widgets.dart';
 
-import '../../../get_instance/src/bindings_interface.dart';
-import '../routes/get_route.dart';
-import '../routes/transitions_type.dart';
+import '../../../../get_instance/src/bindings_interface.dart';
+import 'get_route.dart';
+import '../transitions/transitions_type.dart';
 
 /// Enables the user to customize the intended pop behavior
 ///