+++ lib/get_navigation/src/routes/core/get_route.dart
@@ -6,9 +6,9 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-import '../../../get_instance/src/bindings_interface.dart';
-import '../../../get_state_manager/src/simple/get_state.dart';
-import '../../get_navigation.dart';
+import '../../../../get_instance/src/bindings_interface.dart';
+import '../../../../get_state_manager/src/simple/get_state.dart';
+import '../../../get_navigation.dart';
 
 class GetPage<T> extends Page<T> {
   final GetPageBuilder page;