+++ lib/get_navigation/src/routes/core/route_middleware.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:flutter/cupertino.dart';
 
-import '../../../getxify.dart';
+import '../../../../getxify.dart';
 
 /// The Page Middlewares.
 /// The Functions will be called in this order
@@ -72,7 +72,7 @@ abstract class GetMiddleware {
   /// {@end-tool}
   GetPage? onPageCalled(GetPage? page) => page;
 
-  /// This function will be called right before the [BindingsInterface] are initialize.
+  /// This function will be called right before the [BindingsInterface] are initialized.
   /// Here you can change [BindingsInterface] for this page
   /// {@tool snippet}
   /// ```dart
@@ -87,7 +87,7 @@ abstract class GetMiddleware {
   /// {@end-tool}
   List<R>? onBindingsStart<R>(List<R>? bindings) => bindings;
 
-  /// This function will be called right after the [BindingsInterface] are initialize.
+  /// This function will be called right after the [BindingsInterface] are initialized.
   GetPageBuilder? onPageBuildStart(GetPageBuilder? page) => page;
 
   /// This function will be called right after the
@@ -167,7 +167,7 @@ class PageRedirect {
     this.settings,
   });
 
-  // redirect all pages that needes redirecting
+  // redirect all pages that needs redirecting
   GetPageRoute<T> getPageToRoute<T>(
     GetPage rou,
     GetPage? unk,