+++ lib/get_navigation/src/routes/router/parse_route.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/foundation.dart';
 
-import '../../../getxify.dart';
+import '../../../../getxify.dart';
 
 @immutable
 class RouteDecoder {